Move SMConfig out of util.py into config.py
