+++ metaspace/engine/sm/engine/annotation/formula_centroids.py
@@ -11,7 +11,8 @@ from botocore.exceptions import ClientError
 from pyspark import SparkContext  # pylint: disable=unused-import
 
 from sm.engine.isocalc_wrapper import IsocalcWrapper  # pylint: disable=unused-import
-from sm.engine.util import SMConfig, split_s3_path
+from sm.engine.util import split_s3_path
+from sm.engine.config import SMConfig
 
 logger = logging.getLogger('engine')
