+++ metaspace/engine/sm/engine/annotation_lithops/annotation_job.py
@@ -26,7 +26,8 @@ from sm.engine.es_export import ESExporter
 from sm.engine.image_store import ImageStoreServiceWrapper
 from sm.engine.isocalc_wrapper import IsocalcWrapper
 from sm.engine.molecular_db import read_moldb_file
-from sm.engine.util import SMConfig, split_s3_path, split_cos_path
+from sm.engine.util import split_s3_path, split_cos_path
+from sm.engine.config import SMConfig
 from sm.engine.utils.perf_profile import Profiler
 
 logger = logging.getLogger('engine')
