+++ metaspace/engine/sm/engine/annotation_spark/msm_basic_search.py
@@ -28,7 +28,7 @@ from sm.engine.fdr import FDR
 from sm.engine.formula_parser import safe_generate_ion_formula
 from sm.engine.isocalc_wrapper import IsocalcWrapper
 from sm.engine.molecular_db import MolecularDB
-from sm.engine.util import SMConfig
+from sm.engine.config import SMConfig
 from sm.engine.utils.perf_profile import Profiler
 
 logger = logging.getLogger('engine')
