+++ metaspace/engine/sm/engine/es_export.py
@@ -21,7 +21,8 @@ from sm.engine.formula_parser import format_ion_formula
 from sm.engine.isocalc_wrapper import IsocalcWrapper
 from sm.engine import molecular_db
 from sm.engine.molecular_db import MolecularDB
-from sm.engine.util import SMConfig, retry_on_exception
+from sm.engine.util import retry_on_exception
+from sm.engine.config import SMConfig
 
 logger = logging.getLogger('engine')
