+++ metaspace/engine/sm/engine/postprocessing/colocalization.py
@@ -12,7 +12,7 @@ from sm.engine.annotation_lithops.executor import Executor
 from sm.engine.annotation_lithops.io import save_cobj, iter_cobjs_with_prefetch
 from sm.engine.dataset import Dataset
 from sm.engine.ion_mapping import get_ion_id_mapping
-from sm.engine.util import SMConfig
+from sm.engine.config import SMConfig
 from sm.engine.image_store import ImageStoreServiceWrapper
 
 COLOC_JOB_DEL = 'DELETE FROM graphql.coloc_job WHERE ds_id = %s AND moldb_id = %s'
