+++ metaspace/engine/sm/rest/datasets.py
@@ -8,7 +8,7 @@ from sm.engine.es_export import ESExporter
 from sm.engine.image_store import ImageStoreServiceWrapper
 from sm.engine.queue import QueuePublisher, SM_ANNOTATE, SM_DS_STATUS, SM_UPDATE, SM_LITHOPS
 from sm.engine.errors import UnknownDSID, DSIsBusy
-from sm.engine.util import SMConfig
+from sm.engine.config import SMConfig
 from sm.rest.dataset_manager import SMapiDatasetManager, DatasetActionPriority
 from sm.rest.utils import NOT_EXIST, INTERNAL_ERROR, body_to_json, OK
